Move frame_provider.py
