+++ cvat/apps/dataset_manager/bindings.py
@@ -33,8 +33,8 @@ from django.utils import timezone
 from cvat.apps.dataset_manager.formats.utils import get_label_color
 from cvat.apps.engine import models
 from cvat.apps.engine.cache import MediaCache
-from cvat.apps.engine.frame_provider import FrameOutputType, TaskFrameProvider
 from cvat.apps.engine.lazy_list import LazyList
+from cvat.apps.engine.media_io.frame_provider import FrameOutputType, TaskFrameProvider
 from cvat.apps.engine.model_utils import add_prefetch_fields
 from cvat.apps.engine.models import (
     AttributeSpec,
