+++ cvat/apps/dataset_manager/formats/cvat.py
@@ -40,7 +40,7 @@ from cvat.apps.dataset_manager.bindings import (
     match_dm_item,
 )
 from cvat.apps.dataset_manager.util import make_zip_archive
-from cvat.apps.engine.frame_provider import FrameOutputType, make_frame_provider
+from cvat.apps.engine.media_io.frame_provider import FrameOutputType, make_frame_provider
 from cvat.apps.engine.models import FrameQuality, TaskMode
 
 from .registry import dm_env, exporter, importer
