+++ cvat/apps/engine/cache.py
@@ -911,10 +911,7 @@ class MediaCache:
         # Otherwise we might need to download files.
         # This is not needed for video tasks, as it will reduce performance,
         # because of reading multiple files (chunks)
-        from cvat.apps.engine.media_io.frame_provider import (
-            FrameOutputType,
-            make_frame_provider,
-        )
+        from cvat.apps.engine.media_io.frame_provider import FrameOutputType, make_frame_provider
 
         task_frame_provider = make_frame_provider(db_task)
